reporting/dashboard/src/net/sourceforge/cruisecontrol/dashboard/web/PanopticodeController.java
/********************************************************************************
 * CruiseControl, a Continuous Integration Toolkit
 * Copyright (c) 2007, ThoughtWorks, Inc.
 * 200 E. Randolph, 25th Floor
 * Chicago, IL 60601 USA
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions
 * are met:
 *
 *     + Redistributions of source code must retain the above copyright
 *       notice, this list of conditions and the following disclaimer.
 *
 *     + Redistributions in binary form must reproduce the above
 *       copyright notice, this list of conditions and the following
 *       disclaimer in the documentation and/or other materials provided
 *       with the distribution.
 *
 *     + Neither the name of ThoughtWorks, Inc., CruiseControl, nor the
 *       names of its contributors may be used to endorse or promote
 *       products derived from this software without specific prior
 *       written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE REGENTS OR
 * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
 * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
 * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
 * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
 * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ********************************************************************************/
package net.sourceforge.cruisecontrol.dashboard.web;

import java.io.File;
import java.io.IOException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import net.sourceforge.cruisecontrol.dashboard.Configuration;
import org.apache.commons.io.FileUtils;
import org.apache.commons.lang.StringUtils;
import org.apache.log4j.Logger;
import org.springframework.web.servlet.ModelAndView;
import org.springframework.web.servlet.mvc.Controller;

public class PanopticodeController implements Controller {

    private final Configuration configuration;

    private static final Logger LOGGER = Logger.getLogger(PanopticodeController.class);

    public PanopticodeController(Configuration configuration) {
        this.configuration = configuration;
    }

    public ModelAndView handleRequest(HttpServletRequest request, HttpServletResponse response) throws Exception {
        String[] url = StringUtils.split(request.getRequestURI(), '/');
        String projectName = url[url.length - 2];
        String category = url[url.length - 1];
        File svgFile = getPanopticodeOutput(projectName, category);
        if (!svgFile.exists()) {
            return pictureNotExist(projectName, response);
        }
        response.setContentType("image/svg+xml");
        response.getWriter().write(FileUtils.readFileToString(svgFile, "UTF-8"));
        return null;
    }

    private File getPanopticodeOutput(String projectName, String category) {
        return new File(configuration.getCruiseConfigDirLocation() + "/projects/" + projectName
                + "/target/reports/svg/interactive-" + category + "-treemap.svg");
    }

    private ModelAndView pictureNotExist(String projectName, HttpServletResponse response) {
        response.setContentType("text/plain");
        try {
            response.getWriter().write("No panopticode output found for " + projectName);
        } catch (IOException e) {
            LOGGER.warn(e);
        }
        return null;
    }
}
